Fix that SHIT!
Okay so basically the problem is that there's a giant ass statement
here trying to use the coordinates as a key for a hash of Nodes.
Basically my hacky fix for this HACKATHON that should never be used
irl (and also I don't know python so YOLO)... Too much coffee....

Okay I am rambling, sorry. Anyway the actual fix is that I made a
hash with all of the coordiates as strings storing the "smart"
variables {0, 1, 2} which map to danger, food, other snakes, etc etc.

Then I check the hash at the key of the coordinate and the proper
value is returned so we can use these badass methods without writing
them from scratch!!!! YASSSS BITCH